+++ .gitignore
@@ -5,8 +5,11 @@
 
 # User-specific stuff:
 .idea/**/workspace.xml
-.idea/**/tasks.xml
+.idea/**/tasks.xml :
+
 .idea/dictionaries
+
+.cache
 
 # Sensitive or high-churn files:
 .idea/**/dataSources/
