+++ calenderView/package.json
@@ -13,6 +13,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "mocha": "^8.0.0"
+    "regenerator-runtime": "^0.13.5"
+  },
+  "devDependencies": {
+    "mocha": "^8.0.1"
   }
 }
