+++ calenderView/src/calendar/app.js
@@ -1,13 +1,12 @@
 import 'regenerator-runtime/runtime'
 
-import {CalenderContent} from "./calenderContent.js";
+import {CalenderService} from "./calenderService.js";
 import {CalenderBody} from "./calenderBody.js";
 import {CalendarMapper} from "./calendarMapper.js";
 
 
 
+let service = new CalenderService();
+let mapper = new CalendarMapper(service);
 
-
-let mapper = new CalendarMapper();
-let view = new CalenderContent();
-new CalenderBody(view, mapper)
+new CalenderBody(service, mapper)
