+++ calenderView/src/calendar/calendarMapper.js
@@ -1,23 +1,19 @@
+
+
 export class CalendarMapper {
 
     constructor(){}
 
 
-    calendarMapper(data, user1, user2, user3){
+    calendarMapper(data) {
 
-        let row1 = document.getElementById("user1");
-        let row2 = document.getElementById("user2");
-        let row3 = document.getElementById("user3");
-        let row4 = document.getElementById("user4");
-
-
-        const calenderMap = new Map(); // neue Haupt-Map initialisiert
+        const calenderMap = new Map();
         let mainContainer = [];
         for (let i = 0; i < data.length; i++) {
             mainContainer.push(data[i]);
         }
 
-        let userMap = new Map(); // neue Map initialisiert für ein user mit event
+        let userMap = new Map();
 
         for (let x = 0; x < mainContainer.length; x++) {
 
@@ -56,188 +52,11 @@ export class CalendarMapper {
             }
         }
 
-
-        if (userMap.size > 0) {          // bearbeitung des letzten Eintrags
+        if (userMap.size > 0) {
             const key = mainContainer[mainContainer.length - 1].eventdate;
             calenderMap.set(key, userMap)
         }
 
-        let timeArray = [];
-
-        for (let [key, value] of calenderMap.entries()) {
-
-
-            let r = user1;
-            let s = user2;
-            let f = user3;
-            let z = user4;
-
-            let time = new Date(key);
-            let tag = time.getDate();
-            let xy = value.entries();
-
-            let a = [];
-            let b = [];
-            let c = [];
-            let d = [];
-
-
-            for (let [user, termin] of xy) {
-
-                function pushItems(users) {
-
-                    termin.forEach((val, key)=>{
-                        let appointment = termin[key].description;
-                        let identifier = termin[key].id.id;
-                        let object = {meet: appointment, nr: identifier}
-                        users.push(object)
-                    })
-                }
-
-                if (user === r) {
-                    pushItems(a)
-                }
-                if (user === s) {
-                    pushItems(b)
-                }
-                if (user === f) {
-                    pushItems(c)
-                }
-                if (user === z) {
-                    pushItems(d)
-                }}
-
-
-            timeArray.push(tag);
-
-            let date = new Date();
-            let firstDay = new Date(date.getFullYear(), date.getMonth(), 1);
-            let today = new Date(firstDay);
-            let firstOfMonth = today.getDate();
-            let firstEntry = timeArray[0];
-            let firstEmptyFields = firstEntry - firstOfMonth;
-            let currentDate = timeArray[timeArray.length - 2];
-            let nextDate = timeArray[timeArray.length - 1];
-            let diffTage = nextDate - currentDate;
-
-
-            if (timeArray > 0) {
-                fillIn(firstEmptyFields)
-            }
-
-            if (diffTage > 1) {
-
-                for (let x = 1; x < diffTage; x++) {
-                    let element = createField("", "");
-                    row1.appendChild(element);
-                    let element1 = createField("", "");
-                    row2.appendChild(element1);
-                    let element2 = createField("", "");
-                    row3.appendChild(element2);
-                    let element3 = createField("", "");
-                    row4.appendChild(element3)
-                }
-            }
-
-
-            let a2 = a.map(function(e) {
-                return e.nr
-            });
-            let b2 = b.map(function(e) {
-                return e.nr
-            });
-            let c2 = c.map(function(e) {
-                return e.nr
-            });
-            let d2 = d.map(function(e) {
-                return e.nr
-            });
-
-            let element = createField(a, a2);
-            row1.appendChild(element);
-            let element1 = createField(b, b2);
-            row2.appendChild(element1);
-            let element2 = createField(c, c2);
-            row3.appendChild(element2);
-            let element3 = createField(d, d2);
-            row4.appendChild(element3);
-        }
-
-        let lastDay = new Date();
-        let lastday = new Date(lastDay.getFullYear(), lastDay.getMonth() + 1, 0, 23, 59, 59);
-        let end = new Date(lastday);
-        let lastEnd = end.getDate();
-
-        let lastEntry = timeArray.pop();
-        let fillUp = lastEnd - lastEntry;
-
-
-        if (calenderMap.size === 0) { // Gitter auffüllen mit leeren Feldern falls es im Monat keine Termine hat
-            fillIn(lastEnd)
-        }
-
-        fillIn(fillUp);
-
-        function createField(data, id) {
-
-            const element = document.createElement("li");
-
-            let button = document.createElement("button");
-            button.classList.add("button");
-
-            button.setAttribute("id", id);
-            button.innerText = "Löschen";
-            button.style.backgroundColor = "green";
-
-
-            for (let x = 0; x < id.length; x++){
-
-                button.onclick = function () {
-
-                    fetch('http://localhost:3000/events/' + id[x], {
-                        method: 'DELETE',
-                    })
-                        .then(res => res.json())
-                        .catch((error)=>{
-                            console.log(error)
-                        });
-                    location.reload()
-                };
-            }
-
-
-            if (button.id === ""){
-                button.style.display = "none"
-            }
-            else {
-                let a = [];
-                for (let x =0; x < data.length; x++){
-
-                    a.push(data[x].meet + "<br>");
-                }
-
-                element.innerHTML = a;
-                element.appendChild(button);
-                button.id = id
-            }
-            return element
-        }
-
-
-
-        function fillIn(input) {
-
-            for (let x = 0; x < input; x++) {
-
-                let element = createField("", "");
-                row1.appendChild(element);
-                let element1 = createField( "", "");
-                row2.appendChild(element1);
-                let element2 = createField("", "");
-                row3.appendChild(element2);
-                let element3 = createField("", "");
-                row4.appendChild(element3)
-            }
-        }
+        return calenderMap
     }
 }
