+++ calenderView/src/calendar/calenderService.js
@@ -1,6 +1,6 @@
 
 
-export class CalenderContent {
+export class CalenderService {
 
 
     constructor() {}
@@ -33,7 +33,6 @@ export class CalenderContent {
     }
 
 
-
     get listUser(){
 
         return fetch('http://localhost:3000/names')
@@ -42,7 +41,6 @@ export class CalenderContent {
 
     addNewAppointment(event) {
 
-        // Übermitteln
         fetch('http://localhost:3000/create', {
             method: "POST",
             headers: {'Content-Type': 'application/json'},
@@ -54,5 +52,20 @@ export class CalenderContent {
             .catch (function (err) {
                 console.log('error: ' + err);
             })
+    }
+
+
+    deleteAppointment(id) {
+        for (let x = 0; x < id.length; x++) {
+
+            fetch('http://localhost:3000/events/' + id[x], {
+                method: 'DELETE',
+            })
+                .then(res => res.json())
+                .catch((error) => {
+                    console.log(error)
+                });
+            location.reload()
+        }
     }
 }
